lesson 6 refactoring

+++ src/main/java/com/findme/validation/RelationshipValidationImpl.java
@@ -23,68 +23,55 @@ public class RelationshipValidationImpl implements RelationshipValidation {
                                                FriendRelationshipStatus newStatus, Integer allRequests,
                                                Integer allFriendsUserFrom, Integer allFriendsUserTo) {
 
-        // REQUESTED
-        if (FriendRelationshipStatus.REQUESTED.equals(newStatus)) {
-            return requestValidation(relationship, newStatus, userFrom, userTo, allRequests);
-        }
-
-        // ACCEPTED
-        if (FriendRelationshipStatus.ACCEPTED.equals(newStatus)) {
-            return acceptValidation(relationship, newStatus, userFrom,
+        // CURRENT STATUS - REQUESTED
+        if (FriendRelationshipStatus.REQUESTED.equals(relationship.getFriendRelationshipStatus())) {
+            return changeFromRequestValidation(relationship, newStatus, userFrom, userTo,
                     allFriendsUserFrom, allFriendsUserTo);
         }
 
-        // DELETED
-        if (FriendRelationshipStatus.DELETED.equals(newStatus)) {
+        // CURRENT STATUS - ACCEPTED
+        if (FriendRelationshipStatus.ACCEPTED.equals(relationship.getFriendRelationshipStatus())
+            && FriendRelationshipStatus.DELETED.equals(newStatus)) {
             return deleteValidation(relationship, newStatus);
         }
 
-        // CANCELED
-        if (FriendRelationshipStatus.CANCELED.equals(newStatus)) {
-            return canceledValidation(relationship, newStatus, userFrom);
+        // CURRENT STATUS - DELETED, CANCELED, DENIED, NULL
+        if (FriendRelationshipStatus.DELETED.equals(relationship.getFriendRelationshipStatus())
+            || FriendRelationshipStatus.CANCELED.equals(relationship.getFriendRelationshipStatus())
+            || FriendRelationshipStatus.DENIED.equals(relationship.getFriendRelationshipStatus())
+            || relationship.getFriendRelationshipStatus() == null) {
+            return changeToRequestValidation(relationship, newStatus, userFrom, userTo, allRequests);
         }
-
-        // DENIED
-        if (FriendRelationshipStatus.DENIED.equals(newStatus)) {
-            return deniedValidation(relationship, newStatus, userFrom);
-        }
-        throw new BadRequestException("Unknown relationship status.");
+        throw new BadRequestException("Invalid request.");
     }
 
 
-    // - максимальное ко-во исходящих заявок в друзья для одного пользователя - 10
-    private Relationship requestValidation(Relationship relationship, FriendRelationshipStatus newStatus,
-                                           User userFrom, User userTo, Integer allRequests) {
-        if (FriendRelationshipStatus.ACCEPTED.equals(relationship.getFriendRelationshipStatus())
-        || FriendRelationshipStatus.REQUESTED.equals(relationship.getFriendRelationshipStatus())) {
-            throw new BadRequestException("Invalid request.");
-        }
-        requestQuantityValidation(allRequests);
-        relationship.setUserFrom(userFrom);
-        relationship.setUserTo(userTo);
-        relationship.setFriendRelationshipStatus(newStatus);
-        return relationship;
-    }
 
 
     // - максимальный допустимый список друзей для одного пользователя - 100
-    private Relationship acceptValidation(Relationship relationship, FriendRelationshipStatus newStatus,
-                                          User userFrom, Integer allFriendsUserFrom, Integer allFriendsUserTo) {
+    private Relationship changeFromRequestValidation(Relationship relationship, FriendRelationshipStatus newStatus,
+                                          User userFrom, User userTo, Integer allFriendsUserFrom, Integer allFriendsUserTo) {
         if (relationship.getUserFrom().equals(userFrom)
-                && FriendRelationshipStatus.REQUESTED.equals(relationship.getFriendRelationshipStatus())) {
+                && FriendRelationshipStatus.CANCELED.equals(newStatus)) {
+            relationship.setFriendRelationshipStatus(newStatus);
+            return relationship;
+        } else if (relationship.getUserFrom().equals(userTo)
+                && FriendRelationshipStatus.ACCEPTED.equals(newStatus)) {
            friendQuantityValidation(allFriendsUserFrom, allFriendsUserTo);
             relationship.setFriendRelationshipStatus(newStatus);
             return relationship;
+        } else if (relationship.getUserFrom().equals(userTo)
+                && FriendRelationshipStatus.DENIED.equals(newStatus)) {
+            relationship.setFriendRelationshipStatus(newStatus);
+            return relationship;
         }
-        throw new BadRequestException("You can accept requested status only");
+        throw new BadRequestException("Invalid request");
     }
-
-
 
 
     // удалять с друзей можно если пользователи являются друзьями как минимум 3 дня
     private Relationship deleteValidation(Relationship relationship, FriendRelationshipStatus newStatus) {
-        if (!FriendRelationshipStatus.ACCEPTED.equals(relationship.getFriendRelationshipStatus())) {
+        if (!FriendRelationshipStatus.DELETED.equals(newStatus)) {
             throw new BadRequestException("You can delete accepted status only");
         }
         if (LocalDate.now().minusDays(3).isBefore(relationship.getDateLastUpdated())) {
@@ -94,27 +81,19 @@ public class RelationshipValidationImpl implements RelationshipValidation {
         return relationship;
     }
 
-
-    private Relationship canceledValidation(Relationship relationship, FriendRelationshipStatus newStatus,
-                                            User userFrom) {
-        if (relationship.getUserFrom().equals(userFrom) &&
-                FriendRelationshipStatus.REQUESTED.equals(relationship.getFriendRelationshipStatus())) {
+    // - максимальное ко-во исходящих заявок в друзья для одного пользователя - 10
+    private Relationship changeToRequestValidation(Relationship relationship, FriendRelationshipStatus newStatus,
+                                                   User userFrom, User userTo, Integer allRequests) {
+        if (FriendRelationshipStatus.REQUESTED.equals(newStatus)) {
             relationship.setFriendRelationshipStatus(newStatus);
+            relationship.setUserFrom(userFrom);
+            relationship.setUserTo(userTo);
+            requestQuantityValidation(allRequests);
             return relationship;
         }
-        throw new BadRequestException("You can cancel your requested status only");
+        throw new BadRequestException("Invalid request");
     }
 
-
-    private Relationship deniedValidation(Relationship relationship, FriendRelationshipStatus newStatus,
-                                          User userFrom) {
-        if (relationship.getUserFrom().equals(userFrom)
-                && FriendRelationshipStatus.REQUESTED.equals(relationship.getFriendRelationshipStatus())) {
-            relationship.setFriendRelationshipStatus(newStatus);
-            return relationship;
-        }
-        throw new BadRequestException("You can denied other user requested status only");
-    }
 
 
     private void friendQuantityValidation(Integer allFriendsUserFrom, Integer allFriendsUserTo) {
